requirements.txt
pymongo~=4.1.1
mongomock~=4.0.0
flask~=2.1.0
flask-classy~=0.6.10
flask-cors~=3.0.10
bcrypt~=3.2.0
pyjwt~=2.3.0
python-dotenv
pytest~=7.1.1